Recover hit volume check return type

+++ include/main/objhits.h
@@ -285,7 +285,7 @@ int ObjHits_TestTaperedCapsule3D(float* point, float pointRadius, float baseRadi
 void ObjHits_SortSweepEntries(ObjHitsSweepEntry** sweepPtrs, int entryCount);
 void ObjHits_TickPriorityHitCooldowns(void);
 void ObjHitbox_UpdateRotatedBounds(ObjHitbox* hitbox, int advanceMatrix);
-u8 ObjHits_CheckHitVolumes(int objA, int objB, int srcObj, char checkA, char checkB, u32 mask, u32 volMask);
+int ObjHits_CheckHitVolumes(int objA, int objB, int srcObj, char checkA, char checkB, u32 mask, u32 volMask);
 void doNothing_800333C8(int objA, int objB, int att, void* state, void* attState, f32 dt);
 void ObjHits_CheckObjectHitVolumes(int objA, int objB, int attA, int attB, f32 dt);
 void ObjHits_RegisterActiveHitVolumeObject(GameObject* obj);
